requirements.txt
google-api-python-client==2.107.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
pandas==2.1.1
matplotlib==3.8.0
streamlit==1.27.2 